Basic authorization backend

Creates AuthBackendBasic class object.

Format

R6::R6Class object.

Construction

AuthBackendBasic$new(FUN)

FUN :: function
character(1), character(1) -> logical(1)
Function to perform authentication which takes two arguments - user and password. Returns boolean - whether access is allowed for a requested user or not.

Methods

authenticate(request, response)
Request, Response -> NULL
Provide authentication for the given request.

References

RFC7617 Wikipedia

See also

Examples

# init users database
user_db = list(
  "user-1" = "password-1",
  "user-2" = "password-2"
)
# define authentication handler
auth_fun = function(user, password) {
  if (is.null(user_db[[user]])) return(FALSE) # not found
  if (!identical(user_db[[user]], password)) return(FALSE) # incorrect
  return(TRUE)
}
# init backend
auth_backend = AuthBackendBasic$new(FUN = auth_fun)

# test backend
# define credentials (see RFC)
creds = jsonlite::base64_enc("user-1:password-1")
# generate request headers
h = list("Authorization" = sprintf("Basic %s", creds))
# simulate request
rq = Request$new(path = "/", headers = h)
# init response object
rs = Response$new()
# perform authentication
auth_backend$authenticate(rq, rs) # TRUE

#> [1] TRUE