Creates RestRserveApplication.

Creates RestRserveApplication object. RestRserveApplication facilitates in turning user-supplied R code into high-performance REST API by allowing to easily register R functions for handling http-requests.

RestRserveApplication

Format

R6Class object.

Usage

app = RestRserveApplication$new()

app$add_endpoint(path = "/echo", method = "GET", FUN = function(request) { RestRserve::create_response(payload = request$query[[1]], content_type = "text/plain") })

app$routes()

For usage details see Methods, Arguments and Examples sections.

Methods

$new()

Constructor for RestRserveApplication. For the moment doesn't take any parameters.

$add_endpoint(path, method, FUN)

Adds endpoint and register user-supplied R function as a handler. user function must return object of the class "RestRserveResponse" which can be easily constructed with create_response

$run(port = "80", ...)

starts RestRserve application from current R session. port - http port for application. ... - key-value pairs of the Rserve configuration.

$call_handler(request, path)

Used internally, usually users don't need to call it. Calls handler function for a given request and path.

$routes()

Lists all registered routes

$check_path_exists(path)

Mainly for internal usage. Returns TRUE/FALSE if path registered / not registered

$check_path_method_exists(path, method)

Mainly for internal usage. Returns TRUE/FALSE path-method pair registered / not registered

$print_endpoints_summary()

Prints all the registered routes with allowed methods

$add_openapi(path = "/openapi.yaml", openapi = openapi_create())

Adds endpoint to serve OpenAPI description of available methods.

$add_swagger_ui(path = "/__swagger__/", path_openapi_yaml = "/openapi.yaml")

Adds endpoint to show swagger-ui. After calling it should be available at http://host:port/__swagger__/

Arguments

app

A RestRserveApplication. object

path

character of length 1. Should be valid path for example '/a/b/c'

method

character of length 1. At the moment one of "GET", "POST", "HEAD"

FUN

function which takes exactly one argument - request. request R object returned by RestRserve:::parse_request() function. Object corresponds to http-request and essentially request is a list with a fixed set of fields. Representation of the "GET" request to "http://localhost:8001/somemethod?a=1&b=2" will look like:

uri

= "/somemethod", always character of length 1

method

= "GET", always character of length 1

query

= c("a" = "1", "b" = "2"), character vector

content_type

= "", always character of length 1

content_length

= 0L, always integer of length 1

body

= NULL.

NULL if the http body is empty or zero length.

raw vector with a "content-type" attribute in all cases except URL encoded form (if specified in the headers)

named characeter vector in the case of a URL encoded form. It will have the same shape as the query string (named string vector)

client_ip

= "0.0.0.0", always character of length 1

raw_cookies

= "", always character of length 1

Examples

echo_handler = function(request) {
 RestRserve::create_response(payload = request$query[[1]],
                             content_type = "text/plain",
                            headers = "Location: /echo",
                            status_code = 201L)
}
app = RestRserveApplication$new()
app$add_endpoint(path = "/echo", method = "GET", FUN = echo_handler)

#> [1] TRUE

req = list(query = c("a" = "2"), method = "GET")
answer = app$call_handler(request = req, path = "/echo")
answer$payload

#> [1] "2"

# "2"