starts RestRserve application

Assuming that RserveApplication is deployed to a dir, restrserve_start starts Rserve from it. Application "deployed" means that directory contains Rserve.conf file which has been generated with restrserve_deploy function

restrserve_start(dir, debug = FALSE, ...)

Arguments

dir
character, path to the directory where application was deployed

debug
logical, FALSE by default. Whether to start Rserve in debug mode.

...
other parameters to Rserve call

Value

named integer. Value is a PID of the Rserve daemon. Name is a path to a file where PID is stored. After start Rserve creates a file with its PID. By default tt is called "Rserve.pid" and created inside dir. This could be specified in restrserve_deploy during application deployment.